feature/browse listing/search helpers out of utils into service layer
