+++ templates/browse.html
@@ -16,7 +16,6 @@
     </div>
 
     <div class="mb-6 rounded-xl border border-slate-200 bg-white p-3 shadow-sm">
-      <!-- Single horizontal row; scroll sideways on small screens -->
       <div
         class="browse-toolbar flex flex-nowrap items-center gap-2 overflow-x-auto pb-0.5 [-webkit-overflow-scrolling:touch] sm:gap-3"
         role="toolbar"
